Adding tests for runs
