+++ test_infra/test_files/30001/RUN1/run_metadata.json
@@ -1,0 +1,3 @@
+{
+  "run_name": "RUN1"
+}
