+++ test_infra/test_files/30001/RUN2/run_metadata.json
@@ -1,0 +1,3 @@
+{
+  "run_name": "RUN2"
+}
